sends things to keep
