+++ Service/credentials.json
@@ -1,0 +1,1 @@
+{"web":{"client_id":"979102031910-j88f39hhlpehrfp999nls476b87ud4md.apps.googleusercontent.com","project_id":"learning-api-proj","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://oauth2.googleapis.com/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"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","redirect_uris":["http://localhost:3000/oauth","http://localhost:5000/oauth","http://localhost:8000/oauth"],"javascript_origins":["http://127.0.0.1","https://127.0.0.1","http://localhost","https://localhost"]}}
