+++ Service/token.json
@@ -1,0 +1,1 @@
+{"type":"authorized_user","client_id":"979102031910-j88f39hhlpehrfp999nls476b87ud4md.apps.googleusercontent.com","client_secret":"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"}
